# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/campaigns/reject": {
            "patch": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Admin can Reject Campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Campaign"
                ],
                "summary": "Admin can Reject Campaign",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id ",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.RejectCampaignResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/campaigns/reported/details": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Admin can see details reported post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Reported"
                ],
                "summary": "Admin can see details reported post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID ",
                        "name": "postId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.ReportDetailsResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/dashboard/User": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Admin can see User toplist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Dashboard"
                ],
                "summary": "Admin can see User toplist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.CategoryStatsResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/dashboard/posts": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Admin can see Post toplist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Dashboard"
                ],
                "summary": "Admin can see Post toplist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.PostStatsResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/users/list": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "Admin can see User List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Users"
                ],
                "summary": "Admin can see User List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "searchkey",
                        "name": "searchkey",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.UserListResponse"
                        }
                    }
                }
            }
        },
...
        "/user/notifications/clear": {
            "delete": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "User can Clear Notification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Notifications"
                ],
                "summary": "User can Clear Notification",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.ClearHistoryResponse"
                        }
                    }
                }
            }
        },
        "/user/notifications/delete": {
            "delete": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "User can delete Notification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Notifications"
                ],
                "summary": "User can Delete Notification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification Id",
                        "name": "notificationId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.CommentPostResponse"
                        }
                    }
                }
            }
        },
...
        "/user/post/comment/delete": {
            "delete": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "User can Report post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Post"
                ],
                "summary": "User can report a post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comment ID",
                        "name": "commentId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.CommentPostResponse"
                        }
                    }
                }
            }
        },
...
        "/user/post/donate/history/clear": {
            "delete": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "User can Clear History",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Donations"
                ],
                "summary": "User can Clear History",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.ClearHistoryResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pb.Category": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "categoryid": {
                    "type": "integer"
                }
            }
        },
...
        "pb.CategoryStatsResponse": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Category"
                    }
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "pb.ClearHistoryResponse": {
            "type": "object",
            "properties": {
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "pb.Comment": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "postid": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                },
                "userid": {
                    "type": "integer"
                }
            }
        },
        "pb.CommentPostResponse": {
            "type": "object",
            "properties": {
                "Post": {
                    "$ref": "#/definitions/pb.PostDetails"
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "pb.PostDetails": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Comment"
                    }
                },
                "post": {
                    "$ref": "#/definitions/pkg_user_pb.Post"
                }
            }
        },
...
        "pb.PostStatsResponse": {
            "type": "object",
            "properties": {
                "posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg_admin_pb.Post"
                    }
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "pb.RejectCampaignResponse": {
            "type": "object",
            "properties": {
                "post": {
                    "$ref": "#/definitions/pkg_admin_pb.Post"
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "pb.ReportDetailsResponse": {
            "type": "object",
            "properties": {
                "post": {
                    "$ref": "#/definitions/pb.ReportedPost"
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "pb.ReportedPost": {
            "type": "object",
            "properties": {
                "post": {
                    "$ref": "#/definitions/pkg_admin_pb.Post"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "pb.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "pb.UserListResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "user": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.User"
                    }
                }
            }
        },
...
        "pkg_admin_pb.Post": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "collected": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "place": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "userid": {
                    "type": "integer"
                }
            }
        },
        "pkg_user_pb.Post": {
            "type": "object",
            "properties": {
                "account_no": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "collected": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "likes": {
                    "type": "integer"
                },
                "place": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}