# pkg/admin/routes/UserStats.go
package routes

import (
	"context"
	"errors"
	"hand/pkg/admin/pb"
	"log"
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
)

//edit

// Admin User Stats godoc
//
//	@Summary		Admin can see User toplist
//	@Description	Admin can see User toplist
//	@Tags			Admin Dashboard
//	@Security		api_key
//	@Accept			json
//	@Produce		json
//	@Param			limit	query		string	false	"limit"
//	@Param			page	query		string	false	"Page number"
//	@Success		200		{object}	pb.CategoryStatsResponse
//	@Router			/admin/dashboard/User  [get]
func UserStats(ctx *gin.Context, c pb.AdminServiceClient) {
	log.Println("Initiating AdminDashboard...")

	pageStr := ctx.Query("page")
	if pageStr == "" {
		pageStr = "1"
	}
	page, err := strconv.Atoi(pageStr)
	if err != nil || page < 0 {
		ctx.AbortWithError(http.StatusBadRequest, errors.New("page format invalid"))
		return
	}

	limitStr := ctx.Query("limit")
	if limitStr == "" {
		limitStr = "10"
	}
	limit, err := strconv.Atoi(limitStr)
	if err != nil || limit < 0 {
		ctx.AbortWithError(http.StatusBadRequest, errors.New("limit format invalid"))
		return
	}
	log.Println("Collected data : ", page, limit)
	//edit
	log.Println("Fetching Data...")

	res, err := c.CategoryStats(context.Background(), &pb.CategoryStatsRequest{})

	if err != nil {
		ctx.AbortWithError(http.StatusBadGateway, err)
		return
	}
	log.Println("Recieved data : ", res)

	ctx.JSON(http.StatusOK, &res)
}